Changed defaultForceWeight to 0.25

+++ src/main/java/edu/odu/icat/model/Force.java
@@ -17,7 +17,7 @@ public class Force extends Base {
      */
     Force() {}
 
-    public Force(Entity origin, Entity destination, int weight) {
+    public Force(Entity origin, Entity destination, double weight) {
         setOrigin(origin);
         setDestination(destination);
         setWeight(weight);
